Ported the vector field struct to use hash maps (unordered_map) versus the standard red and black std::map for better cache utilization.
Fixed temporarily issues with size_x_ and size_y_ not being properly assigned in the VectorField by getting the boundary size with the square root of the vector field map size. WIll be fixed later.

Revamped traversal to maintain order by switching foreach functions (as there is no longer a maintained order within the underlying data structure) to nested for loops.

Created a hash struct to allow for proper hashing of the IndexPair key in the unordered map located in the VectorField class.

Modified fluid_sim.cpp procedurally in accordance with the data structure switch.

+++ CUDATest/index_pair.cpp
@@ -9,7 +9,7 @@ bool IndexPair::operator==(const IndexPair& i1) const {
 	return x == i1.x && y == i1.y;
 }
 
-unsigned int IndexPair::operator()(const IndexPair& i1) const noexcept {
+size_t IndexPair::operator()(const IndexPair& i1) const noexcept {
 	unsigned int hash1 = std::hash<unsigned int>()(i1.x);
 	unsigned int hash2 = std::hash<unsigned int>()(i1.y);
 	return hash1 ^ (hash2 << 1);
